+++ include/moves.h
@@ -1,8 +1,8 @@
-unsigned long long posKingMoves(Board* b, int color);
-unsigned long long posKnightMoves(Board* b, int color, int index);
-unsigned long long posPawnMoves(Board* b, int color, int index);
-unsigned long long posRookMoves(Board* b, int color, int index);
-unsigned long long posBishMoves(Board* b, int color, int index);
-unsigned long long posQueenMoves(Board* b, int color, int index);
+uint64_t posKingMoves(Board* b, int color);
+uint64_t posKnightMoves(Board* b, int color, int index);
+uint64_t posPawnMoves(Board* b, int color, int index);
+uint64_t posRookMoves(Board* b, int color, int index);
+uint64_t posBishMoves(Board* b, int color, int index);
+uint64_t posQueenMoves(Board* b, int color, int index);
 
 int isInCheck(Board* b, int kingsColor);
